rename query folder to catalog
