+++ packages/toolkit/src/lib/react-query-service/catalog/useDeleteCatalog.ts
