+++ packages/toolkit/src/lib/react-query-service/catalog/useDeleteCatalogFile.tsx
