+++ packages/toolkit/src/lib/react-query-service/catalog/useGetFileDetails.ts
