+++ packages/toolkit/src/lib/react-query-service/catalog/useUpdateCatalog.ts
