+++ packages/toolkit/src/view/knowledge/components/ChunkCard.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 
 import { Icons, Separator, Switch, Tag } from "@instill-ai/design-system";
 
-import { Chunk } from "../../../lib/react-query-service/knowledge/types";
+import { Chunk } from "../../../lib/react-query-service/catalog/types";
 
 type ChunkCardProps = {
   chunk: Chunk;
