+++ packages/toolkit/src/view/knowledge/components/FileTable.tsx
@@ -5,7 +5,7 @@ import * as React from "react";
 import {
   File,
   FileStatus,
-} from "../../../lib/react-query-service/knowledge/types";
+} from "../../../lib/react-query-service/catalog/types";
 import { FileTableHeader } from "./FileTableHeader";
 import { FileTableRow } from "./FileTableRow";
 import { getStatusSortValue } from "./lib/helpers";
