+++ packages/toolkit/src/view/knowledge/components/FileTableRow.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 
 import { Button, Dialog, Icons, Tag } from "@instill-ai/design-system";
 
-import { File } from "../../../lib/react-query-service/knowledge/types";
+import { File } from "../../../lib/react-query-service/catalog/types";
 import {
   convertFileType,
   formatDate,
