+++ packages/toolkit/src/view/knowledge/components/tabs/MarkdownTab.tsx
@@ -12,7 +12,7 @@ import {
   Textarea,
 } from "@instill-ai/design-system";
 
-import { KnowledgeBase } from "../../../../lib/react-query-service/knowledge/types";
+import { KnowledgeBase } from "../../../../lib/react-query-service/catalog/types";
 
 type MarkdownTabProps = {
   knowledgeBase: KnowledgeBase;
@@ -123,7 +123,7 @@ export const MarkdownTab = ({ knowledgeBase }: MarkdownTabProps) => {
                     onClick={() => requestSort("fileName")}
                   >
                     {sortConfig.key === "fileName" &&
-                    sortConfig.direction === "ascending" ? (
+                      sortConfig.direction === "ascending" ? (
                       <Icons.ChevronUp className="h-4 w-4 stroke-semantic-fg-secondary" />
                     ) : (
                       <Icons.ChevronDown className="h-4 w-4 stroke-semantic-fg-secondary" />
@@ -156,7 +156,7 @@ export const MarkdownTab = ({ knowledgeBase }: MarkdownTabProps) => {
                     onClick={() => requestSort("fileType")}
                   >
                     {sortConfig.key === "fileType" &&
-                    sortConfig.direction === "ascending" ? (
+                      sortConfig.direction === "ascending" ? (
                       <Icons.ChevronUp className="h-4 w-4 stroke-semantic-fg-secondary" />
                     ) : (
                       <Icons.ChevronDown className="h-4 w-4 stroke-semantic-fg-secondary" />
@@ -189,7 +189,7 @@ export const MarkdownTab = ({ knowledgeBase }: MarkdownTabProps) => {
                     onClick={() => requestSort("processedStatus")}
                   >
                     {sortConfig.key === "processedStatus" &&
-                    sortConfig.direction === "ascending" ? (
+                      sortConfig.direction === "ascending" ? (
                       <Icons.ChevronUp className="h-4 w-4 stroke-semantic-fg-secondary" />
                     ) : (
                       <Icons.ChevronDown className="h-4 w-4 stroke-semantic-fg-secondary" />
@@ -222,7 +222,7 @@ export const MarkdownTab = ({ knowledgeBase }: MarkdownTabProps) => {
                     onClick={() => requestSort("createTime")}
                   >
                     {sortConfig.key === "createTime" &&
-                    sortConfig.direction === "ascending" ? (
+                      sortConfig.direction === "ascending" ? (
                       <Icons.ChevronUp className="h-4 w-4 stroke-semantic-fg-secondary" />
                     ) : (
                       <Icons.ChevronDown className="h-4 w-4 stroke-semantic-fg-secondary" />
@@ -255,7 +255,7 @@ export const MarkdownTab = ({ knowledgeBase }: MarkdownTabProps) => {
                     onClick={() => requestSort("createTime")}
                   >
                     {sortConfig.key === "createTime" &&
-                    sortConfig.direction === "ascending" ? (
+                      sortConfig.direction === "ascending" ? (
                       <Icons.ChevronUp className="h-4 w-4 stroke-semantic-fg-secondary" />
                     ) : (
                       <Icons.ChevronDown className="h-4 w-4 stroke-semantic-fg-secondary" />
@@ -267,7 +267,7 @@ export const MarkdownTab = ({ knowledgeBase }: MarkdownTabProps) => {
                 <div key={index}>
                   <div className="h-16 flex-col items-center justify-center px-3">
                     <div className=" product-label-label-2">Retrievable</div>
-                    <Switch checked={item.status} onCheckedChange={() => {}} />
+                    <Switch checked={item.status} onCheckedChange={() => { }} />
                   </div>
                   {index !== sortedData.length - 1 ? (
                     <Separator orientation="horizontal" />
